server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram (>=3.20.0.post0,<4.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "tortoise-orm (>=0.25.0,<0.26.0)",
    "aerich (>=0.8.2,<0.9.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "tomlkit (>=0.13.2,<0.14.0)",
    "fastapi (>=0.115.12,<0.116.0)",

    "ffmpeg-python (>=0.2.0,<0.3.0)",

    "python-multipart (>=0.0.20,<0.0.21)",
]

[[project.authors]]
name = "Your Name"
email = "you@example.com"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.aerich]
tortoise_orm = "config_reader.TORTOISE_ORM"
location = "./db/migrations"
src_folder = "./."
